TestCellEdgeSrn replaced by TestCellEdgeInteriorSrn
The new test incorporates  the edge-only case as well the case with both edge and interior SRN
