+++ cell_based/src/odes/DeltaNotchEdgeOdeSystem.cpp
@@ -55,8 +55,9 @@ DeltaNotchEdgeOdeSystem::DeltaNotchEdgeOdeSystem(std::vector<double> stateVariab
     SetDefaultInitialCondition(1, 1.0); // soon overwritten
 
     this->mParameters.push_back(0.5);
-    //By default zero. If no interior SRN model is specified, interior delta is zero
-    this->mParameters.push_back(0);
+    //By default zero. If no interior SRN model is specified, interior delta/notch is zero
+    this->mParameters.push_back(0.0);
+    this->mParameters.push_back(0.0);
     if (stateVariables != std::vector<double>())
     {
         SetStateVariables(stateVariables);
@@ -69,13 +70,14 @@ DeltaNotchEdgeOdeSystem::~DeltaNotchEdgeOdeSystem()
 
 void DeltaNotchEdgeOdeSystem::EvaluateYDerivatives(double time, const std::vector<double>& rY, std::vector<double>& rDY)
 {
-    double notch = rY[0];
-    double delta = rY[1];
-    double neigh_delta = this->mParameters[0]; // Shorthand for "this->mParameter("neighbor delta");"
-
+    const double notch = rY[0];
+    const double delta = rY[1];
+    const double neigh_delta = this->mParameters[0]; // Shorthand for "this->mParameter("neighbor delta");"
+    const double interior_delta = this->mParameters[1];
+    const double interior_notch = this->mParameters[2];
     // The next two lines define the ODE system by Collier et al. (1996)
-    rDY[0] = neigh_delta*neigh_delta/(0.01 + neigh_delta*neigh_delta) - notch;  // d[Notch]/dt
-    rDY[1] = 1.0/(1.0 + 100.0*notch*notch) - delta;                   // d[Delta]/dt
+    rDY[0] = neigh_delta*neigh_delta/(0.01 + neigh_delta*neigh_delta) - notch+0.1*interior_notch;  // d[Notch]/dt
+    rDY[1] = 1.0/(1.0 + 100.0*notch*notch) - delta+0.1*interior_delta;// d[Delta]/dt
 }
 
 template<>
@@ -93,7 +95,8 @@ void CellwiseOdeSystemInformation<DeltaNotchEdgeOdeSystem>::Initialise()
     this->mParameterUnits.push_back("non-dim");
     this->mParameterNames.push_back("interior delta");
     this->mParameterUnits.push_back("non-dim");
-
+    this->mParameterNames.push_back("interior notch");
+    this->mParameterUnits.push_back("non-dim");
     this->mInitialised = true;
 }
