+++ cell_based/src/simulation/modifiers/DeltaNotchEdgeInteriorTrackingModifier.cpp
@@ -84,18 +84,21 @@ void DeltaNotchEdgeInteriorTrackingModifier<DIM>::UpdateCellData(AbstractCellPop
         const double this_notch = p_interior_model->GetNotch();
         const double this_delta = p_interior_model->GetDelta();
         double total_edge_delta = 0;
+        double total_edge_notch = 0;
         const unsigned int n_cell_edges = p_cell_edge_model->GetNumEdgeSrn();
         std::vector<double> edge_delta(n_cell_edges);
         for (unsigned i = 0 ; i  < p_cell_edge_model->GetNumEdgeSrn(); i++)
         {
             boost::shared_ptr<DeltaNotchSrnEdgeModel> p_model
             = boost::static_pointer_cast<DeltaNotchSrnEdgeModel>(p_cell_edge_model->GetEdgeSrn(i));
-            total_edge_delta += p_model->GetDelta();
+            total_edge_delta += p_model->GetNeighbouringDelta();
+            total_edge_notch += p_model->GetNotch();
         }
         // Note that the state variables must be in the same order as listed in DeltaNotchOdeSystem
         cell_iter->GetCellData()->SetItem("interior notch", this_notch);
         cell_iter->GetCellData()->SetItem("interior delta", this_delta);
-        cell_iter->GetCellData()->SetItem("total edge delta", total_edge_delta);
+        cell_iter->GetCellData()->SetItem("total neighbour edge delta", total_edge_delta);
+        cell_iter->GetCellData()->SetItem("total edge notch", total_edge_notch);
     }
 
 }
